requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0
beautifulsoup4
scrapingbee
readability-lxml
openai>=1.0.0
langchain-core>=0.1.0
langchain>=0.1.0
langchain-community